Render windows etc, then write everything at once & bugfix
Instead of recursively writing to terminal for Frame, Window, Minibuffer, etc,
first render the lines of strings to write and their positions, then
write everything at together in TedTui::redraw.
This separation of concerns allows for removal of all `should_redraw` members
from the many affected structures. The new logic for calculating what to draw
and what to keep should result in less data written to the terminal.
However, because of the need to render the image in order to see if anything has changed,
the editor has slowed down somewhat, at least in debug build.

Also, fix bug where multi-char graphemes where treated as multiple
sigils when moving and deleting