Add text prompt. M-g for go to line
Prompts exist right under the bottom modeline. They prompt for text input,
and on submit, i.e. <enter>, the input text is sent to a callback
to be parsed and perform some action.
E.g. M-g, go to line, asks for a line number, and on submit,
parses an int from the input string and jumps to that line